+++ src/tightdb/index_string.cpp
@@ -24,7 +24,7 @@ Array* StringIndex::create_node(Allocator& alloc, bool is_leaf)
 {
     Array::Type type = is_leaf ? Array::type_HasRefs : Array::type_InnerBptreeNode;
     UniquePtr<Array> top(new Array(alloc)); // Throws
-    top->create(type); // Throws
+    top->create(type); // Throws LEAK
 
     // Mark that this is part of index
     // (as opposed to columns under leaves)
@@ -32,7 +32,7 @@ Array* StringIndex::create_node(Allocator& alloc, bool is_leaf)
 
     // Add subcolumns for leaves
     Array values(alloc);
-    values.create(Array::type_Normal); // Throws
+    values.create(Array::type_Normal); // Throws LEAK
     values.ensure_minimum_width(0x7FFFFFFF); // This ensures 31 bits plus a sign bit
     top->add(values.get_ref()); // first entry in refs points to offsets
 
@@ -54,7 +54,6 @@ StringIndex::key_type StringIndex::GetLastKey() const
     get_child(*m_array, 0, offsets);
     return key_type(offsets.back());
 }
-
 
 
 void StringIndex::insert_with_offset(size_t row_ndx, StringData value, size_t offset)
@@ -369,7 +368,7 @@ bool StringIndex::LeafInsert(size_t row_ndx, key_type key, size_t offset, String
         StringData v2 = get(row_ndx2, buffer);
         if (v2 == value) {
             if (m_deny_duplicate_values)
-                throw UniqueConstraintViolation();
+                throw LogicError(LogicError::unique_constraint_violation);
             // convert to list (in sorted order)
             Array row_list(alloc);
             row_list.create(Array::type_Normal); // Throws
@@ -399,7 +398,7 @@ bool StringIndex::LeafInsert(size_t row_ndx, key_type key, size_t offset, String
         StringData v2 = get(r1, buffer);
         if (v2 == value) {
             if (m_deny_duplicate_values)
-                throw UniqueConstraintViolation();
+                throw LogicError(LogicError::unique_constraint_violation);
             // find insert position (the list has to be kept in sorted order)
             // In most cases we refs will be added to the end. So we test for that
             // first to see if we can avoid the binary search for insert position