+++ src/tightdb/table.cpp
@@ -1274,6 +1274,13 @@ void Table::add_search_index(size_t col_ndx)
         index.set_parent(&m_columns, column_pos+1);
         index_ref = index.get_ref();
     }
+    else if (type == col_type_Int) {
+        Column& col = get_column(col_ndx);
+        col.create_search_index(); // Throws
+        StringIndex& index = *static_cast<StringIndex*>(col.m_search_index);
+        index.set_parent(&m_columns, column_pos + 1);
+        index_ref = index.get_ref();
+    }
     else {
         throw LogicError(LogicError::illegal_combination);
     }
@@ -1307,7 +1314,7 @@ bool Table::has_primary_key() const TIGHTDB_NOEXCEPT
 }
 
 
-void Table::add_primary_key(size_t col_ndx)
+bool Table::try_add_primary_key(size_t col_ndx)
 {
     if (TIGHTDB_UNLIKELY(!is_attached()))
         throw LogicError(LogicError::detached_accessor);
@@ -1332,14 +1339,21 @@ void Table::add_primary_key(size_t col_ndx)
         AdaptiveStringColumn& col_2 = static_cast<AdaptiveStringColumn&>(col);
         StringIndex& index = col_2.get_search_index();
         if (index.has_duplicate_values())
-            throw UniqueConstraintViolation();
+            return false;
         index.set_allow_duplicate_values(false);
     }
     else if (type == col_type_StringEnum) {
         ColumnStringEnum& col_2 = static_cast<ColumnStringEnum&>(col);
         StringIndex& index = col_2.get_search_index();
         if (index.has_duplicate_values())
-            throw UniqueConstraintViolation();
+            return false;
+        index.set_allow_duplicate_values(false);
+    }
+    else if (type == col_type_Int) {
+        Column& col_2 = static_cast<Column&>(col);
+        StringIndex& index = *static_cast<StringIndex*>(col_2.get_search_index());
+        if (index.has_duplicate_values())
+            return false;
         index.set_allow_duplicate_values(false);
     }
     else {
@@ -1349,13 +1363,15 @@ void Table::add_primary_key(size_t col_ndx)
     }
 
     int attr = m_spec.get_column_attr(col_ndx);
-    attr |= col_attr_PrimaryKey;
+    attr |= col_attr_Unique | col_attr_PrimaryKey;
     m_spec.set_column_attr(col_ndx, ColumnAttr(attr)); // Throws
 
 #ifdef TIGHTDB_ENABLE_REPLICATION
     if (Replication* repl = get_repl())
         repl->add_primary_key(this, col_ndx); // Throws
 #endif
+
+    return true;
 }
 
 
@@ -1371,7 +1387,7 @@ void Table::remove_primary_key()
     for (size_t col_ndx = 0; col_ndx < num_cols; ++col_ndx) {
         int attr = m_spec.get_column_attr(col_ndx);
         if (attr & col_attr_PrimaryKey) {
-            attr &= ~col_attr_PrimaryKey;
+            attr &= ~(col_attr_Unique | col_attr_PrimaryKey);
             m_spec.set_column_attr(col_ndx, ColumnAttr(attr)); // Throws
             m_primary_key = 0;
 
@@ -2890,7 +2906,7 @@ size_t Table::do_find_pkey_string(StringData value) const
 }
 
 
-template <class T> size_t Table::find_first(size_t col_ndx, T value) const
+template<class T> size_t Table::find_first(size_t col_ndx, T value) const
 {
     TIGHTDB_ASSERT(!m_columns.is_attached() || col_ndx < m_columns.size());
     TIGHTDB_ASSERT(get_real_column_type(col_ndx) == ColumnTypeTraits3<T>::ct_id_real);
@@ -3608,7 +3624,7 @@ public:
             for (size_t i = 0; i != n; ++i) {
                 int attr = spec.get_column_attr(i);
                 // Remove any index specifying attributes
-                attr &= ~(col_attr_Indexed | col_attr_PrimaryKey);
+                attr &= ~(col_attr_Indexed | col_attr_Unique | col_attr_PrimaryKey);
                 spec.set_column_attr(i, ColumnAttr(attr)); // Throws
             }
             size_t pos = spec.m_top.write(out); // Throws
@@ -3712,8 +3728,8 @@ void Table::update_from_parent(size_t old_baseline) TIGHTDB_NOEXCEPT
 
 
 // to JSON: ------------------------------------------
-void Table::to_json_row(std::size_t row_ndx, std::ostream& out, size_t link_depth, std::map<std::string, 
-    std::string>* renames) const
+void Table::to_json_row(std::size_t row_ndx, std::ostream& out, size_t link_depth, std::map<std::string,
+                        std::string>* renames) const
 {
     std::map<std::string, std::string> renames2;
     renames = renames ? renames : &renames2;